+++ org.kde.plasma.persian-calendar/contents/ui/GeneralConfig.qml
@@ -1,46 +1,54 @@
-import QtQuick 2.0
-import QtQuick.Layouts 1.1
-import QtQuick.Controls 1.3
+import QtQuick 2.12
+import QtQuick.Layouts 1.11
+import QtQuick.Controls 2.12
 import org.kde.plasma.plasmoid 2.0
+import org.kde.kirigami as Kirigami
 import org.kde.plasma.core 2.0 as PlasmaCore
 import org.kde.plasma.components 3.0 as PlasmaComponents
 
 Item
 {
+  id: configRoot
+
+  signal configurationChanged
+
   property alias cfg_widgetWidth:    widgetWidth.value
   property alias cfg_showTooltip:    showTooltip.checked
   property alias cfg_updateInterval: updateInterval.value
   property alias cfg_mainText:       mainText.text
   property alias cfg_tooltipText:    tooltipText.text
+  
+  property alias cfg_persianDateLabel: persianDateLabel.height
 
-  anchors.centerIn:     parent
+  anchors.centerIn: parent
 
   ColumnLayout
   {
-
+    /*
     RowLayout
     {
       Label { text: "Widget width" }
       SpinBox
       {
         id:           widgetWidth
-        minimumValue: 10
-        maximumValue: 10000
-        decimals:     0
+        from:         10
+        to:           10000
+        editable:     true
         stepSize:     10
-        suffix:       ""
+        // suffix:       ""
       }
-    }
+    } */
 
     RowLayout
     {
       CheckBox
       {
         id: showTooltip
+        checked: true
         text: "Show tooltip"
       }
     }
-
+    /*
     RowLayout
     {
       Label { text: "Update Interval (seconds)" }
@@ -53,8 +61,8 @@ Item
         stepSize:     1
         suffix:       " s"
       }
-    }
-
+    } */
+    /*
     RowLayout
     {
       Label { text: "Main" }
@@ -65,8 +73,8 @@ Item
         placeholderText:  "HTML & CSS are supported"
         text:             ""
       }
-    }
-
+    } */
+    /*
     RowLayout
     {
       Label { text: "Tooltip" }
@@ -77,6 +85,6 @@ Item
         placeholderText:  "HTML & CSS are supported"
         text:             ""
       }
-    }
+    } */
   }
 }